perf(gen-text): re-architecture to use generators for sentences on the fly, print progress, magic num refactor
Replace the generate in place -> sentence list architecture with a sentence generator.
This massively reduces the amount of memory consumed and creates a dramatic speedup due
to the reduction in memory management overhead.

Replace default min, and max character code magic numbers with constants
defined at the top of the script.
Print progress every 10 percent when the --file option is used.